% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_tabla_indicadores.R
\name{crear_tabla_indicadores}
\alias{crear_tabla_indicadores}
\title{Creates and export excel with count by achieved indicator}
\usage{
crear_tabla_indicadores(db = data_c, file_xlsx, exit_dir)
}
\arguments{
\item{db}{A tibble with the indicators to be presented in the table.  This is the output
of \code{create_tibble_count()}}

\item{exit_dir}{A string with the name of the excel.xlsx file to be crated}
}
\value{
an excel sheet stored in exit_dir/file_xlsx
}
\description{
Creates and export excel with count by achieved indicator
}
