% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reports.R
\name{create_reports}
\alias{create_reports}
\title{Create reports for dashboards and tables}
\usage{
create_reports(
  dir_downloads = "downloads",
  dir_clean = "clean_data",
  reports = c("Download_outputs", "all_projects", "All_Quarters", "expected_outputs",
    "expected_int", "expected_outcomes", "KPIs_Report"),
 
    refresh_token = "1000.b11df28b89daaeb2df10fa2c43178db6.6f953944b607f0ff366915cb9a770edc",
  ...
)
}
\arguments{
\item{dir_downloads}{A directory where downloaded reports will be stored}

\item{dir_clean}{A directory where the clean reports will be stored
keep only trainings -----------------------------
name indicator as they are expressed in the logframe
drop training unclean indicators
bind with training indicators
All of these results were reported
results reported
join with expected
clean indicators for expected results
set count to 0 if the result was expected
identify if the result was expected
Define elements of the logframe
transform type to ordered factor ----------------------------------
clean Indicator text -------------------------------------------
clean status for those that are expected but not reported yet
categorise indicators based on the elements of the logframe
sort elements ------------------------------------------------------}
}
\description{
This function downloads the reports from Zoho, stores the raw data in downloads
and saves the clean data in clean data
}
