% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_time.R
\name{criteria_time}
\alias{criteria_time}
\title{Define criteria to dowload records modified since the last download}
\usage{
criteria_time(reference_data)
}
\arguments{
\item{reference_data}{A tible. Last download of the report}
}
\value{
A string to be passed to the query that downloads the reports from zoho. "Modified_Time>'10-Apr-2021 11:37:16'"
}
\description{
Gets the last time a record was modified in the existing data and
creates the criteria to donwload records after that time. The search criteria
is based on: \href{https://www.zoho.com/creator/help/api/v2/get-records.html}{Defining the search criteria}
}
