% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_report.R
\name{download_report}
\alias{download_report}
\title{Downloads report from zoho}
\usage{
download_report(dir_downloads, zoho_report, access_token, ...)
}
\arguments{
\item{dir_downloads}{A directory. Directory to store the download data}

\item{zoho_report}{A string. Report link name as it is defined in Zoho}

\item{access_token}{A string. The refresh token that belongs to the requesting user,
which is returned when the access_type=offline is included in the query string of the authorization request
See \href{https://www.zoho.com/creator/help/api/v2/refresh-the-access-token.html}{Refresh the Access Token}}
}
\description{
Checks when the latest records in downloads was modified and exports
records from zoho that have been modified since that date
}
