% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_pact_token.R
\name{refresh_pact_token}
\alias{refresh_pact_token}
\title{Refreshes the token for zoho uk-pact}
\usage{
refresh_pact_token(refresh_token)
}
\arguments{
\item{refresh_token}{A string. Token that allows refreshing the access_token}
}
\description{
Refreshes the token for zoho uk-pact
}
